+++ desktop/src/main/java/org/vorthmann/zome/app/impl/TrackballRenderingViewer.java
@@ -1,8 +1,8 @@
 package org.vorthmann.zome.app.impl;
 
-import javax.vecmath.Matrix3d;
-import javax.vecmath.Matrix4d;
-import javax.vecmath.Vector3d;
+import javax.vecmath.Matrix3f;
+import javax.vecmath.Matrix4f;
+import javax.vecmath.Vector3f;
 
 import com.vzome.core.viewing.Camera;
 import com.vzome.desktop.controller.CameraController;
@@ -19,14 +19,14 @@ public class TrackballRenderingViewer implements CameraController.Viewer
 {
 	private final CameraController.Viewer delegate;
 	
-	private final Vector3d translation;
+	private final Vector3f translation;
 
 	public TrackballRenderingViewer( CameraController.Viewer delegate )
 	{
 		this .delegate = delegate;
 		
-		this .translation = new Vector3d();
-		Matrix4d matrix = new Matrix4d();
+		this .translation = new Vector3f();
+		Matrix4f matrix = new Matrix4f();
 		Camera defaultCamera = new Camera();
 		defaultCamera .setMagnification( 1.0f );
 		defaultCamera .getViewTransform( matrix );
@@ -40,12 +40,12 @@ public class TrackballRenderingViewer implements CameraController.Viewer
 	}
 
 	@Override
-	public void setViewTransformation( Matrix4d trans )
+	public void setViewTransformation( Matrix4f trans )
 	{
-	    Matrix3d justRotation3d = new Matrix3d();
+	    Matrix3f justRotation3d = new Matrix3f();
 	    trans .get( justRotation3d );
 	    justRotation3d .invert(); // to match the invert() in the caller
-	    Matrix4d finalTransform = new Matrix4d();
+	    Matrix4f finalTransform = new Matrix4f();
 	    finalTransform .set( this .translation );
 	    finalTransform .setRotation( justRotation3d );
 	    finalTransform .invert(); // to match the invert() in the caller
